refactor(server): split cloud mobility store package
