+++ scripts/max_lines_allowlist.txt
@@ -28,7 +28,6 @@ apps/desktop/src/lib/workflows/mcp/connector-persistence.test.ts 704 existing de
 server/proliferate/auth/identity/service.py 807 existing auth service exceeds repo max-lines threshold
 server/proliferate/db/store/automations.py 816 server structure migration debt for oversized store file
 server/proliferate/db/store/billing.py 2605 server structure migration debt for oversized store file
-server/proliferate/db/store/cloud_mobility.py 1410 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_repo_config.py 719 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_sync/commands.py 1565 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_sync/events.py 1109 server structure migration debt for oversized store file
