+++ server/proliferate/db/store/cloud_mobility/__init__.py
@@ -1,0 +1,1 @@
+"""Cloud workspace mobility store package."""
