+++ server/tests/unit/test_cloud_mobility_service_lifecycle.py
@@ -6,7 +6,7 @@ from uuid import uuid4
 
 import pytest
 
-from proliferate.db.store.cloud_mobility import (
+from proliferate.db.store.cloud_mobility.records import (
     CloudWorkspaceHandoffOpValue,
     CloudWorkspaceMobilityValue,
 )
